refactor: one constant per concept, one wraparound idiom
Two consistency gaps that made the code harder to read than it is.

The display message size was the literal 64 in ten places across five
files - MessageConfig's five fields, OverlayState::message,
DisplaySettings::finalMessage, BoundaryCue::message, a local in
config_api.cpp - plus a 63 buried inside copyMessage's strncpy. The exactly
parallel concept, a sound name, already had kSoundNameLength enforced
across the packer, the config and the API. Messages now use
kDisplayMessageLength the same way, and copyMessage takes a
reference-to-array so the size is part of the type and a caller cannot
hand it a shorter buffer.

Deadline comparisons used two different wraparound idioms side by side:
unsigned subtraction in DisplayScheduler::shouldRender, and a signed cast
of the difference in toggleBlinkIfDue, toggleColonIfDue, overlayExpired,
the paged-overlay timer and the RTC health poll. Both survive a millis()
rollover, so having both meant a reader had to work out each time which one
was load-bearing. Standardized on unsigned subtraction, which is correct
without raising the signed-overflow question at all. Signed arithmetic on
unixtime() differences is a different thing and stays signed.

+++ src/clock_application.cpp
@@ -184,8 +184,7 @@ void ClockApplication::processButtonEvents() {
 }
 
 void ClockApplication::checkRtcHealth(uint32_t nowMs) {
-  if (static_cast<long>(nowMs - lastRtcHealthCheckMs_) <
-      static_cast<long>(kRtcHealthPollIntervalMs)) return;
+  if ((nowMs - lastRtcHealthCheckMs_) < kRtcHealthPollIntervalMs) return;
   lastRtcHealthCheckMs_ = nowMs;
   const bool healthy = rtc_.isHealthy();
   if (!healthy) {
